Update Navbar and Home Page layout
The commit includes changes to the Navbar and index (Home Page) of the website. For the Navbar, the padding was updated, and the logo source and dimensions were changed. On the Home Page, the paragraph structure was cleaned up and the Facebook Feed element was rearranged for a better layout.